requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6 